Refactoring ApplicationRunnerMain code in preparation for fast samza sql.
