+++ samza-core/src/main/java/org/apache/samza/runtime/ApplicationRunnerMain.java
@@ -21,10 +21,9 @@ package org.apache.samza.runtime;
 
 import joptsimple.OptionSet;
 import joptsimple.OptionSpec;
-import org.apache.samza.application.ApplicationUtil;
 import org.apache.samza.config.Config;
 import org.apache.samza.util.CommandLine;
-import org.apache.samza.util.Util;
+
 
 /**
  * This class contains the main() method used by run-app.sh.
@@ -51,28 +50,6 @@ public class ApplicationRunnerMain {
     OptionSet options = cmdLine.parser().parse(args);
     Config orgConfig = cmdLine.loadConfig(options);
     ApplicationRunnerOperation op = cmdLine.getOperation(options);
-    invokeApplicationRunner(orgConfig, op);
-  }
-
-  public static ApplicationRunner invokeApplicationRunner(Config orgConfig, ApplicationRunnerOperation op) {
-    Config config = Util.rewriteConfig(orgConfig);
-
-    ApplicationRunner appRunner =
-        ApplicationRunners.getApplicationRunner(ApplicationUtil.fromConfig(config), config);
-
-    switch (op) {
-      case RUN:
-        appRunner.run(null);
-        break;
-      case KILL:
-        appRunner.kill();
-        break;
-      case STATUS:
-        System.out.println(appRunner.status());
-        break;
-      default:
-        throw new IllegalArgumentException("Unrecognized operation: " + op);
-    }
-    return appRunner;
+    ApplicationRunnerUtil.invokeApplicationRunner(orgConfig, op);
   }
 }
